pyproject.toml
[project]
name = "tax-process-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
